Update tinyopt to v1.0.
